packages/llmterface_gemini/pyproject.toml
[project]
name = "llmterface-gemini"
version = "0.1.0"
description = "Gemini provider for LLMterface"
authors = [{ name = "D. Zachary Wheeler", email = "celestialswashbuckler@gmail.com" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
keywords = [
  "llm",
  "ai",
  "gemini",
  "google",
  "provider",
  "plugin"
]
requires-python = ">=3.11,<4.0"
dependencies = [
  "llmterface>=0.1.0,<1.0.0",
  "google-genai>=1.25.0,<2.0.0",
]

[project.entry-points."llmterface.providers"]
gemini = "llmterface_gemini.plugin:PROVIDER"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmterface_gemini"]
